Revert

+++ include/sparseir/basis.hpp
@@ -227,7 +227,7 @@ inline Eigen::VectorXd default_sampling_points(const PiecewiseLegendrePolyVector
     if (u.xmin() != -1.0 || u.xmax() != 1.0)
         throw std::runtime_error("Expecting unscaled functions here.");
 
-    if (static_cast<size_t>(L) < u.size()) {
+    if (L < u.size()) {
         return u.polyvec[L].roots();
     } else {
         // Approximate roots by extrema
